backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
